move members components to the members_modal folder and reference from createDirectChannel component
